% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScyllaCharacterization.R
\docType{package}
\name{ScyllaCharacterization-package}
\alias{ScyllaCharacterization}
\alias{ScyllaCharacterization-package}
\title{ScyllaCharacterization: Characterize cohorts receiving COVID-19 treatments}
\description{
Characterize cohorts receiving COVID-19 treatments
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ohdsi-studies/Scylla/characterization}
  \item Report bugs at \url{https://github.com/OHDSI/ScyllaCharacterization/issues}
}

}
\keyword{internal}
